requirements.txt
asgiref==3.9.1
Django==5.2.5
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-nested-routers==0.94.2
pillow==11.3.0
PyJWT==2.10.1
setuptools==80.9.0
sqlparse==0.5.3
